+++ tsdb/engine/tsm1/engine.go
@@ -1283,6 +1283,15 @@ func (e *Engine) createCallIterator(measurement string, call *influxql.Call, opt
 		return nil, err
 	}
 
+	// Reverse the tag sets if we are ordering by descending.
+	if !opt.Ascending {
+		a := make([]*influxql.TagSet, len(tagSets))
+		for i, t := range tagSets {
+			a[i] = t.Reverse()
+		}
+		tagSets = a
+	}
+
 	// Calculate tag sets and apply SLIMIT/SOFFSET.
 	tagSets = influxql.LimitTagSets(tagSets, opt.SLimit, opt.SOffset)
 
@@ -1334,6 +1343,15 @@ func (e *Engine) createVarRefIterator(measurement string, opt influxql.IteratorO
 	tagSets, err := mm.TagSets(e.id, opt.Dimensions, opt.Condition)
 	if err != nil {
 		return nil, err
+	}
+
+	// Reverse the tag sets if we are ordering by descending.
+	if !opt.Ascending {
+		a := make([]*influxql.TagSet, len(tagSets))
+		for i, t := range tagSets {
+			a[i] = t.Reverse()
+		}
+		tagSets = a
 	}
 
 	// Calculate tag sets and apply SLIMIT/SOFFSET.